1 - Inclusão do mapa do Google da Dashboard; 2 - Criadas todas as telas de manutenção; 3 - Acerto na tela de Serviços; 4 - Criada a abertura de chamados;

+++ SGA/Models/Cargos/Cargo.cs
@@ -9,7 +9,7 @@ namespace SGA.Models.Cargos
     {
         public int Id { get; set; }
         public string CargoDesc { get; set; }
-        public double Salario { get; set; }
+        public decimal Salario { get; set; }
         public int Status { get; set; }
     }
 }
